# Avoid using syncvar_t in the qthreads shim